Move DAG-related facilities to `analyzer` module
